Add patch support
This is not yet complete; what remains is the ability to generate
the final patch, the one that patches patchlevel.h (and thus
lists the locally applied patches in perl -V / Config.pm).

This requires downloading the proper patchlevel.h for each version
and generating a patch in generate.pl.