Add -Duseshrplib to build flags to build a shared libperl.so
Perl's ./Configure can be asked to build a shared library, libperl.so,
which is required by some modules from CPAN (like Imagemagick), or when
you want to embed the Perl interpreter in your own programs.

However, because there may be a performance penalty, the "old" static
builds have not been removed.

More information can be found in the installation manual at
https://metacpan.org/pod/distribution/perl/INSTALL#Building-a-shared-Perl-library