Reduces image size of resulting Docker Perl image
By using a multi-stage strategy the resulting image can be brought down
to ~100MB instead of ~800MB.

By running the following commands we can create two images, a build
image and a slim image:

    docker build -t perl:<tag> --target build .

    docker build -t perl:<tag>-slim .

This can be further optimized by only installing cpanm in another stage.
It is a nice to have while building an image, but you don't need cpanm
on a running container.