generate: Add checksum validation for SSL modules
cpanm defaults to downloading via http, so currently while there is a
checksum for App::cpanminus there isn't a checksum or validation for the
SSL modules installed.

Rather than just switching to HTTPS, explicitly download the modules and
verify the checksums, so the Docker image is built with known versions
and checksums.