Don't run tests with cpanm installs for the image
cpanm runs tests for everything it installs, which might be a good thing
or a bad thing. It depends mostly on flavor. cpm will install modules
without running tests by default for example. This commits still uses
cpanm, but with the --notest flag so we can improve on our build times.

Closes: #195

Signed-off-by: Wesley Schwengle <waterkip@cpan.org>